rename loadingactivity into appcheckactivity
