+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             android:exported="false"
             android:label="HomeActivity" />
         <activity
-            android:name=".views.splash.AppCheckActivity"
+            android:name=".views.app_check.AppCheckActivity"
             android:exported="false" />
         <activity
             android:name=".views.on_boarding.OnBoardingActivity"
