+++ app/src/main/java/com/nutricatch/dev/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.nutricatch.dev.databinding.ActivityMainBinding
-import com.nutricatch.dev.views.splash.AppCheckActivity
+import com.nutricatch.dev.views.app_check.AppCheckActivity
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
